Fix crash when no upstream servers defined.
This is a regession introduced in 3b6df06fb8cb3652d2e7afd085fae3f416408013.

When dnsmasq is started without upstreams (yet), but a
DNS query comes in that needs forwarding dnsmasq now potentially crashes as
the value for "first" variable is undetermined.

A segmentation violation occurs when the index
is out of bounds of  serverarray.

Credits go to pedro0311 <pedro@freshtomato.org>